requirements.txt
# Основные зависимости для криптосистемы Эль-Гамаля
# Python 3.7+ требуется

# GUI библиотека (входит в стандартную библиотеку Python)
# tkinter - уже включен в Python

# Для работы с изображениями
Pillow>=8.0.0

# Для работы с документами Word
python-docx>=0.8.11

# Для типизации (Python 3.5+)
typing-extensions>=3.7.4; python_version < "3.8"
